Accessed API and changed bookCoverURL prop
Introduced an if statement at the start of my render method that
returns null if books2 is empty. This allows ComponentDidMount to
run after the app has rendered for the first time.
I also changed Book so that it now contains an img tag within the book-cover
div. This displays the book cover image without the need for passing the
bookCoverURL prop with a 'url()' enclosing it.