Remove bugs from Book and Bookshelf
Changed the way changeShelf accesses books in books2, so that it
now uses the unique id rather than their index. Bookshelf now uses the
book id as the key for each list item.
Book now changes shelf in the correct way, and also defaults the
select shelf option to the current shelf of the book.