Fixed the SearchBooks component.
Made changes to the SearchBooks component so that it can use the books
contained within books2 rather than books. We still have an issue
with both uses of Book taking a really clunky prop of bookAuthor.
I think the next step is to clean up some of this messy prop passing.