Changed how authors is being passed as a prop.
Dealt with the bug with authors being sorted in Book as opposed to
Bookshelf or SearchBooks. The main problem was that Book was
being called elsewhere in the app, with the dummy books being
passed to it. The props being passed to Book by Bookshelf and
SearchBooks have now been tidied up nicely.