Fixed the image bug
Have made a pretty dodgy fix to the problem of the search function
crashing if there's no image to access, where the Book component takes
entry.imageLinks rather than entry.imageLinks.thumbnail as the
bookCoverURL prop, and the Book prop checks whether entry.imageLinks is
undefined. If it is, bookCoverURL is set to null, if not, bookCoverURL
is set to bookCoverURL.thumbnail.