Made changes to changeShelf
Removed the anti-pattern in changeShelf where we were making
direct changes to the state. We're now using .filter to produce a copy
of this.state.books, and then making changes to that. I've also
changed the parameters being passed to changeShelf, so that now
it's only being passed the event and the book. BookID is no longer
passed, and instead I've replaced it with book.id within the function,
which is much simpler.